LandingView: as dir, moved children to modules
